+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This changelog covers all three packages, as they are (for now) updated as a who
 ## Unrealesed
 
 - Add Share settings screen where you can see & edit rights / access control #113
+- Add Invite form #45
 - Convert Classes to typescript interfaces. Show button for this in Class view. #118
 - `Create new resource` button on Drive
 - Show multiple parents in breadcrumbs